+++ CHANGELOG.md
@@ -32,6 +32,12 @@ For support you can use [the contact form](https://contact.faircode.eu/?product=
 
 ### [Barsboldia](https://en.wikipedia.org/wiki/Barsboldia)
 
+### Next version
+
+* Added experimental force light menu item / button for reformatted message view
+* Small improvements and minor bug fixes
+* Updated [translations](https://crowdin.com/project/open-source-email)
+
 ### 1.2253 2024-12-23
 
 * Fixed showing S/MIME digest algorithm